youtube-final/backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
yt-dlp>=2023.10.13
pydub>=0.25.1
openai-whisper>=20230314
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.4.2
pillow>=10.0.0
matplotlib>=3.7.0
wordcloud>=1.9.0
requests>=2.28.0
numpy>=1.24.0
fpdf==1.7.2
motor==3.1.1
fastapi==0.95.0
yt-dlp==2023.3.4
pillow==9.5.0
matplotlib==3.7.1
wordcloud==1.8.2.2
whisper==1.1.10
google-generativeai==0.3.1
